refac(ImageFeatures) Reorganized so that "image features" like delete, archive, and share (and someday tag management) can be shared between ImageView and MainView.  Put together some dependency injection for DeleteImagesButton for the different contexts (this wasn't necessary for archive and share yet because they haven't really been implemented, they're dummy modules for now).
