+++ Frontend/src/ImageFeatures/DeleteImagesButton.js
@@ -1,8 +1,7 @@
-import React, { useContext } from 'react';
+import React from 'react';
 import { IconButton } from "@mui/material";
 import DeleteIcon from "@mui/icons-material/Delete";
 
-import AppDataContext from '../SupportingModules/AppDataContext';
 import filterSocket from '../SupportingModules/FilterSocket';
 
 
@@ -12,17 +11,16 @@ import filterSocket from '../SupportingModules/FilterSocket';
  * @returns The DeleteImageButton component to be rendered in the app.
  */
 export default function DeleteImagesButton (props) {
-  const { appState } = useContext(AppDataContext);
 
   const handleClick = () => {
     // Inform the backend of the deletion
-    console.log(props.imageIds);
     filterSocket.sendMessage({
       'type': 'deleteImages',
       'imageIds': props.imageIds
     });
-    appState.setSelectedItems([]);
-    appState.setSelectionActive(false);
+    if (props.handleClickInContext) {
+      props.handleClickInContext();
+    };
   };
 
   return(
